+++ src-ui/app/services/mqtt/integrations/hmd-data.mqtt-integration.ts
@@ -27,18 +27,9 @@ export class HMDDataMqttIntegrationService {
       value: 'off',
       available: false,
     });
-    await this.mqtt.initProperty({
-      type: 'SENSOR',
-      id: 'debugHmdActivity',
-      topicPath: 'debugHmdActivity',
-      displayName: 'VR HMD Activity (DEBUG)',
-      value: 'null',
-      available: false,
-    });
     this.openvr.status.subscribe((status) => {
       this.mqtt.setPropertyAvailability('hmdModel', status === 'INITIALIZED');
       this.mqtt.setPropertyAvailability('hmdOnHead', status === 'INITIALIZED');
-      this.mqtt.setPropertyAvailability('debugHmdActivity', status === 'INITIALIZED');
     });
     this.openvr.devices
       .pipe(map((devices) => devices.find((d) => d.class === 'HMD')))
@@ -47,7 +38,6 @@ export class HMDDataMqttIntegrationService {
           [device?.manufacturerName, device?.modelNumber].filter(Boolean).join(' ') ?? 'null';
         this.mqtt.setSensorPropertyValue('hmdModel', name);
         this.mqtt.setSensorPropertyValue('hmdOnHead', device?.hmdOnHead ? 'on' : 'off');
-        this.mqtt.setSensorPropertyValue('debugHmdActivity', device?.debugHmdActivity ?? 'null');
       });
   }
 }