+++ src-ui/app/services/shutdown-automations.service.ts
@@ -56,6 +56,7 @@ export class ShutdownAutomationsService {
   );
   private sleepMode = false;
   private sleepModeLastSet = 0;
+  private triggeredThisSleep = false;
   private aloneSince = 0;
   private isAlone = false;
   private wasNotAlone = false;
@@ -105,7 +106,7 @@ export class ShutdownAutomationsService {
     this.automationConfigService.configs
       .pipe(
         map((configs) => configs.SHUTDOWN_AUTOMATIONS),
-        // Reset the 'last set' in case any of the trigger parameters change, so the user doesn't get any unwanted surprises
+        // Reset the 'alone since' in case any of the trigger parameters change, so the user doesn't get any unwanted surprises
         pairwise(),
         filter(([oldConfig, newConfig]) => {
           const keys: Array<keyof ShutdownAutomationsConfig> = [
@@ -125,6 +126,7 @@ export class ShutdownAutomationsService {
     // Track sleep mode being set
     this.sleepService.mode.pipe(distinctUntilChanged()).subscribe((mode) => {
       this.sleepMode = mode;
+      if (!this.sleepMode) this.triggeredThisSleep = false;
       this.sleepModeLastSet = Date.now();
     });
     // Track being alone
@@ -197,11 +199,18 @@ export class ShutdownAutomationsService {
   private async handleTriggerOnSleep() {
     interval(1000)
       .pipe(
-        filter(() => this._stage.value === 'IDLE'),
+        // Only trigger if this trigger is enabled
         filter(() => this.config.triggersEnabled),
         filter(() => this.config.triggerOnSleep),
+        // Only trigger if the shutdown sequence isn't running
+        filter(() => this._stage.value === 'IDLE'),
+        // Only trigger if the sleep mode is active
         filter(() => this.sleepMode),
+        // Only trigger if we haven't already triggered this sleep (resets once sleep mode disables)
+        filter(() => !this.triggeredThisSleep),
+        // Only trigger if the sleep mode has been active for long enough
         filter(() => Date.now() - this.sleepModeLastSet >= this.config.triggerOnSleepDuration),
+        // Only trigger if we're in the activation window, if it's configured
         filter(
           () =>
             !this.config.triggerOnSleepActivationWindow ||
@@ -209,11 +218,12 @@ export class ShutdownAutomationsService {
               this.config.triggerOnSleepActivationWindowStart,
               this.config.triggerOnSleepActivationWindowEnd
             )
-        ),
-        // Only trigger once every 5 minutes at most
-        throttleTime(300000, asyncScheduler, { leading: true, trailing: false })
+        )
       )
-      .subscribe(() => this.runSequence('SLEEP_TRIGGER'));
+      .subscribe(() => {
+        this.triggeredThisSleep = true;
+        this.runSequence('SLEEP_TRIGGER');
+      });
   }
 
   private async handleTriggerWhenAlone() {